requirements.txt
#Base
Flask

#SQLAlchemy
Flask-SQLAlchemy
flask-restless
flask-bcrypt

#Auth
pyjwt
